refactor: change location & add component container
agregado unalto maximo para los modales con tabla adentro, ademas se le pusu un scroll personalizado dentro de las tablas
