+++ src/components/modals/historys/history services/index.tsx
@@ -1,8 +1,9 @@
 'use client';
 
-import { LayoutModal } from '..';
-import {  Servicio } from '@/types';
+import { LayoutModal } from '../..';
+import { Servicio } from '@/types';
 import { TableHistoryServices } from '@/components/tables/in modals/history services';
+import { ContainerTableHistory } from '..';
 
 export const ModalHistoryServices = ({
     servicios,
@@ -16,9 +17,10 @@ export const ModalHistoryServices = ({
             footer={false}
             isOpen={true}
         >
-            <div className="m-auto mb-4">
+            <ContainerTableHistory>
                 <TableHistoryServices servicios={servicios} />
-            </div>
+            </ContainerTableHistory>
+            
         </LayoutModal>
     );
 };
